Project setup (#1)
Setting up an open-source Kotlin library to hold the Prosemirror code we
implemented during the Native Editor project.
I am using IntelliJ IDEA 2024.1.4 (Community Edition) to build and run
the project. I am unsure if it will work with VsCode.

This is the initial commit and the repo is not public yet. I am planning
to make it public after I finish some to-dos and finish the pipeline
implementation to deploy the library on our public maven repository

Tasks covered in this PR:

- Copying and adapting code from MobileKit repo
- Adding license files (CREDO process - OSR-1000)
- Initial Github actions setup
- Initial maven repository publish setup (need to test)


Co-authored-by: Douglas Marques <dmarques@atlassian.com>