Bump prosemirror-model to 1.22.3 (#28)
Most of the original code was from v1.18.1, but FromDom and ToDom was
from a newer version. So I have updated the whole model module to
v1.22.3.

I haven't converted all the tests in v1.22.3 yet. They are commented out
in the code, so that I can convert them later.

This PR has a breaking change: `ParseRuleImpl` has been replaced with
`TagParseRuleImpl` and `StyleParseRuleImpl`. This change was made in
v1.20.0.