Automatically tag release branch (#43)
### Problem
Originally, `publish_ios` was triggered by the tag, and creates the
release (using the tag), so that it can upload the ios assets to the
release.
`publish.yml` was triggered by a push to release branch because it
uploads to Artifactory, which only allows uploads from the release
branch.
Now that I've changed both to trigger on a push to release branch, the
ios one is failing because there is no tag.

### Changes
I've merged `publish_ios.yml` and `publish.yml` into 1 file. When there
is a push to release branch, it will automatically tag the commit,
publish to artifactory and upload assets to the newly created Github
release.

Thanks to @tchernykh for all the ideas/help!

### Testing
I've tested the workflow (without the Artifcatory publishing, because
that's restricted to `release` branch) on my branch and it works as
expected.