ME-6962: Fixed node.copy returning node with identical NodeID (#87)
Since we converted from ts to kt and added the concept of NodeID, the
existing logic of `if (content == this.content) return this` is no
longer correct as it'll "copy" a node but return a node with the same
NodeID, causing various problems. We need to do a proper copy every time
(thus generating a new NodeID).

Since the overload already covers this (but without the problematic
shortcut), it's fine to just delete this.